Refactor payroll form boundary
